requirements.txt
tkinter
opencv-python
fer
pygame
pillow
